refactor : 파일 구조 변경, css 파일 분리
