+++ src/components/notice/Layout.tsx
@@ -1,6 +1,5 @@
 import TopBar from "./TopBar";
 import TopTab from "./TopTab";
-import NoticeList from "./NoticeList";
 import {useEffect} from "react";
