+++ src/components/notice/TopTab.tsx
@@ -1,10 +1,9 @@
 
 import {useState} from "react";
 import NoticeList from "./NoticeList";
-import styled from '@emotion/styled';
 /** @jsxImportSource @emotion/react */
 import { jsx, css } from '@emotion/react';
-
+import * as styles from '../../styles/constants/TopTab.style'
 const TopTab = ()=>{
   const data = [
     {
@@ -46,8 +45,8 @@ const TopTab = ()=>{
   return(
 
     <div>
-      <Rectangle242/>
-      <Group278>
+      <styles.Rectangle242/>
+      <styles.Group278>
         {data.map((element,idx)=>(
           <div 
             css={[
@@ -57,9 +56,9 @@ const TopTab = ()=>{
               : css`border-bottom: 2px solid #D9D9D9;`,
             ]}
             onClick={() => selectMenuHandler(idx)} >
-            <TabFontBox>
-              <TabFont>{element.title}</TabFont>
-            </TabFontBox>
+            <styles.TabFontBox>
+              <styles.TabFont>{element.title}</styles.TabFont>
+            </styles.TabFontBox>
           </div>
           // <div className={[element.pos,idx === index?styles.focused:styles.not_focused].join(" ")}
           // onClick={()=>selectMenuHandler(idx)}>
@@ -68,7 +67,7 @@ const TopTab = ()=>{
           //   </TabFontBox>
           // </div>
         ))}
-      </Group278>
+      </styles.Group278>
       <p style={{top:"200px", position:"absolute"}}>{data[index].contents}</p>
     </div>
     
@@ -76,67 +75,5 @@ const TopTab = ()=>{
 };
 
 
-`
-/* 공지사항 */
-margin: auto;
-/* Body M 16 */
-font-family: 'Pretendard';
-font-style: normal;
-font-weight: 500;
-font-size: 13px;
-line-height: 24px;
-/* identical to box height, or 24px */
-letter-spacing: -0.011em;
 
-/* black */
-color: #313131;
-
-`;
-const TabFont = styled.div 
-`
-margin: auto;
-/* Body M 16 */
-font-family: 'Pretendard';
-font-style: normal;
-font-weight: 500;
-font-size: 13px;
-line-height: 24px;
-/* identical to box height, or 24px */
-letter-spacing: -0.011em;
-
-/* black */
-color: #313131;
-
-`;
-const TabFontBox = styled.div
-`
-position: absolute;
-width:55px;
-height:24px;
-top:8px;
-left:8px;
-
-`;
-
-const Rectangle242 = styled.div
-`
-box-sizing: border-box;
-position: absolute;
-width: 390px;
-height: 36px;
-left: 0px;
-top: 107px;
-/* white */
-background: #FFFFFF;
-/* gray 3 */
-border-bottom: 2px solid #D9D9D9;
-`;
-const Group278 = styled.div
-`
-position: absolute;
-width: 390px;
-height: 36px;
-left: 0px;
-top: 107px;
-`;
 export default TopTab;
