+++ src/styles/constants/NoticeList.style.ts
@@ -1,0 +1,14 @@
+
+import styled from '@emotion/styled';
+
+export const NoticeListBox = styled.div
+`
+width:390px;
+height:486px;
+overflow : auto;
+-ms-overflow-style: none;
+scrollbar-width: none;
+::webkit-scrollbar{
+  display:none;
+}
+`;
