+++ src/styles/constants/TopTab.style.ts
@@ -1,0 +1,52 @@
+import styled from '@emotion/styled';
+
+export const TabFont = styled.div 
+`
+margin: auto;
+/* Body M 16 */
+font-family: 'Pretendard';
+font-style: normal;
+font-weight: 500;
+font-size: 13px;
+line-height: 24px;
+/* identical to box height, or 24px */
+letter-spacing: -0.011em;
+
+/* black */
+color: #313131;
+
+`;
+
+export const TabFontBox = styled.div
+`
+position: absolute;
+width:55px;
+height:24px;
+top:8px;
+left:8px;
+
+`;
+
+export const Rectangle242 = styled.div
+`
+box-sizing: border-box;
+position: absolute;
+width: 390px;
+height: 36px;
+left: 0px;
+top: 107px;
+/* white */
+background: #FFFFFF;
+/* gray 3 */
+border-bottom: 2px solid #D9D9D9;
+`;
+
+
+export const Group278 = styled.div
+`
+position: absolute;
+width: 390px;
+height: 36px;
+left: 0px;
+top: 107px;
+`;
